requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic>=2.8.0
pypdf>=5.0.0
numpy>=1.26.0
faiss-cpu>=1.8.0.post1
sentence-transformers>=3.0.1
transformers>=4.44.0
torch>=2.4.0
google-generativeai>=0.8.0
